Fix: Tablas arregladas

+++ src/main/java/desarrollo_web_rre/desarrollo_web_rre/models/AvisoRepository.java
@@ -6,8 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvisoRepository extends JpaRepository<Aviso, Long> {
+public interface AvisoRepository extends JpaRepository<Aviso, Integer> {
 
-    // Equivalente a get_avisos: ordenado por fechaIngreso desc y paginado
     Page<Aviso> findAllByOrderByFechaIngresoDesc(Pageable pageable);
 }
